Moved memory mapped file instance to PotreeData, and all JSON reading to Potree2ReaderBase.
